Refactor spell parsing
Break up giant spell parsing javascript
