+++ src/parser/index.js
@@ -2,14 +2,15 @@ import getCharacter from "./character/index.js";
 import getActions from "./features/actions.js";
 import getFeatures from "./features/features.js";
 import getClasses from "./classes/index.js";
-import { default as getSpells, parseItemSpells as getItemSpells } from "./spells/index.js";
+import { getCharacterSpells } from "./spells/getCharacterSpells.js";
+import { getItemSpells } from "./spells/getItemSpells.js";
 import getInventory from "./inventory/index.js";
 
 let parseJson = (ddb) => {
   let character = getCharacter(ddb);
   let features = getFeatures(ddb, character);
   let classes = getClasses(ddb);
-  let spells = getSpells(ddb, character);
+  let spells = getCharacterSpells(ddb, character);
   let actions = getActions(ddb, character);
   let itemSpells = getItemSpells(ddb, character);
   let inventory = getInventory(ddb, character, itemSpells);
