--- src/parser/spells/index.js
@@ -1,1103 +1,0 @@
-import DICTIONARY from "../dictionary.js";
-import utils from "../../utils.js";
-
-let getComponents = (data) => {
-  return {
-    value: data.definition.componentsDescription,
-    vocal: data.definition.components.includes(1),
-    somatic: data.definition.components.includes(2),
-    material: data.definition.components.includes(3),
-    ritual: data.definition.ritual,
-    concentration: data.definition.concentration,
-  };
-};
-
-let getMaterials = (data) => {
-  // this is mainly guessing
-  if (data.definition.componentsDescription && data.definition.componentsDescription.length > 0) {
-    let cost = 0;
-    let matches = data.definition.componentsDescription.toLowerCase().match(/([\d.]+)\s*gp/);
-    if (matches) {
-      cost = parseInt(matches[1].replace("."));
-    }
-
-    return {
-      value: data.definition.componentsDescription,
-      consumed: data.definition.componentsDescription.toLowerCase().indexOf("consume") !== -1,
-      cost: cost,
-      supply: 0,
-    };
-  } else {
-    return {
-      value: "",
-      consumed: false,
-      cost: 0,
-      supply: 0,
-    };
-  }
-};
-
-/**
- * Retrieves the spell preparation mode, depending on the location this spell came from
- *
- */
-let getSpellPreparationMode = (data) => {
-  // default values
-  let prepMode = "prepared";
-  // If always prepared mark as such, if not then check to see if prepared
-  let prepared = data.alwaysPrepared || data.prepared;
-  // handle classSpells
-  if (data.flags.vtta.dndbeyond.lookup === "classSpell") {
-    const classPrepMode = utils.findByProperty(
-      DICTIONARY.spell.preparationModes,
-      "name",
-      data.flags.vtta.dndbeyond.class
-    ).value;
-    if (data.alwaysPrepared) {
-      prepMode = "always";
-    } else if (prepMode) {
-      prepMode = classPrepMode;
-    }
-    // Warlocks should use Pact spells
-    // but lets mark level 0 as regular spells so they show up as cantrips
-    if (prepMode === "pact" && data.definition.level === 0) {
-      prepMode = "prepared";
-      prepared = true;
-    }
-  } else if (data.flags.vtta.dndbeyond.lookup === "race" && data.definition.level !== 0) {
-    // set race spells as innate
-    prepMode = "innate";
-  } else if (
-    // Warlock Mystic Arcanum are passed in as Features
-    data.flags.vtta.dndbeyond.lookupName.startsWith("Mystic Arcanum")
-  ) {
-    // these have limited uses (set with getUses())
-    prepMode = "pact";
-    prepared = false;
-  } else if (data.flags.vtta.dndbeyond.look === "item " && data.definition.level !== 0) {
-    prepared = false;
-    prepMode = "prepared";
-  } else {
-    // If spell doesn't use a spell slot and is not a cantrip, mark as always preped
-    let always = !data.usesSpellSlot && data.definition.level !== 0;
-    let ritaulOnly = data.ritualCastingType !== null || data.castOnlyAsRitual; // e.g. Book of ancient secrets & totem barb
-    if (always && ritaulOnly) {
-      // in this case we want the spell to appear in the spell list unprepared
-      prepared = false;
-    } else if (always) {
-      // these spells are always prepared, and have a limited use that's
-      // picked up by getUses() later
-      // this was changed to "atwill"
-      prepMode = "atwill";
-    }
-    if (data.flags.vtta.dndbeyond.lookup === "classFeature") {
-      if (data.alwaysPrepared) {
-        prepMode = "always";
-      }
-    }
-  }
-
-  return {
-    mode: prepMode,
-    prepared: prepared,
-  };
-};
-
-/**
- * Get the reset condition of the spell, if uses restricted
- * @param {*} data Spell data
- */
-let getUses = (data) => {
-  let resetType = null;
-  let limitedUse = null;
-  // we check this, as things like items have useage attached to the item, not spell
-  if (data.flags.vtta.dndbeyond.limitedUse !== undefined && data.flags.vtta.dndbeyond.limitedUse !== null) {
-    limitedUse = data.flags.vtta.dndbeyond.limitedUse;
-    resetType = DICTIONARY.resets.find((reset) => reset.id == limitedUse.resetType);
-  } else if (data.limitedUse !== undefined && data.limitedUse !== null) {
-    limitedUse = data.limitedUse;
-    resetType = DICTIONARY.resets.find((reset) => reset.id == limitedUse.resetType);
-  }
-
-  if (resetType !== null && resetType !== undefined) {
-    return {
-      value: limitedUse.numberUsed ? limitedUse.maxUses - limitedUse.numberUsed : limitedUse.maxUses,
-      max: limitedUse.maxUses,
-      per: resetType.value,
-    };
-  } else {
-    return {};
-  }
-};
-
-/**
- * Gets the activation information of this spell
- */
-let getActivation = (data) => {
-  let activationType = DICTIONARY.spell.activationTypes.find(
-    (type) => type.activationType === data.definition.activation.activationType
-  );
-  if (activationType && data.definition.activation.activationTime) {
-    return {
-      type: activationType.value,
-      cost: data.definition.activation.activationTime,
-      condition: "",
-    };
-  } else {
-    return {
-      type: "action",
-      cost: 1,
-      condition: "",
-    };
-  }
-};
-
-/**
- * Retrieves the spell duration
- */
-let getDuration = (data) => {
-  if (data.definition.duration) {
-    let units = "";
-    if (data.definition.duration.durationUnit !== null) {
-      units = data.definition.duration.durationUnit.toLowerCase();
-    } else {
-      units = data.definition.duration.durationType.toLowerCase().substring(0, 4);
-    }
-    return {
-      value: data.definition.duration.durationInterval || "",
-      units: units,
-    };
-  } else {
-    return {};
-  }
-};
-
-/**
- * Does the spell target creatures?
- * @param {*} data
- */
-let doesTargetCreature = (data) => {
-  const creature = /a creature you|creature( that)? you can see|interrupt a creature|would strike a creature|creature of your choice|creature or object within range|cause a creature|creature must be within range/gi;
-  const creaturesRange = /(humanoid|monster|creature|target)(s)? (or loose object )?(of your choice )?(that )?(you can see )?within range/gi;
-  return data.definition.description.match(creature) || data.definition.description.match(creaturesRange);
-};
-
-/**
- * Get Target Values
- * Uses regex magic to try and determine the number of creatures affected
- * @param {*} data
- */
-let getTargetValues = (data) => {
-  const numCreatures = /(?!At Higher Levels.*)(\w*) (falling )?(willing )?(creature|target|monster|celestial|fiend|fey|corpse(s)? of|humanoid)(?!.*you have animated)/gim;
-  const targets = [...data.definition.description.matchAll(numCreatures)];
-  const targetValues = targets
-    .filter((target) => {
-      const matches = DICTIONARY.numbers.filter((n) => n.natural === target[1].toLowerCase());
-      return Array.isArray(matches) && !!matches.length;
-    })
-    .map((target) => DICTIONARY.numbers.find((n) => n.natural === target[1].toLowerCase()).num);
-
-  if (Array.isArray(targetValues) && !!targetValues.length) {
-    return Math.max(...targetValues);
-  } else {
-    return null;
-  }
-};
-
-/**
- * Spell targets
- */
-let getTarget = (data) => {
-  // if spell is an AOE effect get some details
-  if (data.definition.range.aoeType && data.definition.range.aoeValue) {
-    return {
-      value: data.definition.range.aoeValue,
-      type: data.definition.range.aoeType.toLowerCase(),
-      units: "ft",
-    };
-  }
-
-  // else lets try and fill in some target details
-  let type = null;
-  let units = null;
-  let value = null;
-
-  // does the spell target a creature?
-  const creatures = doesTargetCreature(data);
-
-  if (creatures) {
-    value = getTargetValues(data);
-  }
-
-  switch (data.definition.range.origin) {
-    case "Touch":
-      units = "touch";
-      if (creatures) type = "creature";
-      break;
-    case "Self":
-      type = "self";
-      break;
-    case "None":
-      type = "none";
-      break;
-    case "Ranged":
-      if (creatures) type = "creature";
-      break;
-    case "Feet":
-      if (creatures) type = "creature";
-      break;
-    case "Miles":
-      if (creatures) type = "creature";
-      break;
-    case "Sight":
-    case "Special":
-      units = "special";
-      break;
-    case "Any":
-      units = "any";
-      break;
-    case undefined:
-      type = null;
-      break;
-    // no default
-  }
-
-  return {
-    value: value,
-    units: units,
-    type: type,
-  };
-};
-
-/** Spell range */
-let getRange = (data) => {
-  // else lets try and fill in some target details
-  let value = data.definition.range.rangeValue ? data.definition.range.rangeValue : null;
-  let units = "ft";
-  let long = null;
-
-  switch (data.definition.range.origin) {
-    case "Touch":
-      value = null;
-      units = "touch";
-      break;
-    case "Self":
-      value = null;
-      units = "self";
-      break;
-    case "None":
-      units = "none";
-      break;
-    case "Ranged":
-      units = "ft";
-      break;
-    case "Feet":
-      units = "ft";
-      break;
-    case "Miles":
-      units = "ml";
-      break;
-    case "Sight":
-    case "Special":
-      units = "special";
-      break;
-    case "Any":
-      units = "any";
-      break;
-    case undefined:
-      units = null;
-      break;
-    // no default
-  }
-
-  return {
-    value: value,
-    long: long,
-    units: units,
-  };
-};
-
-let getActionType = (data) => {
-  if (data.definition.requiresSavingThrow && !data.definition.requiresAttackRoll) {
-    return "save";
-  }
-
-  if (
-    data.definition.tags.includes("Damage") &&
-    data.definition.range.rangeValue &&
-    data.definition.range.rangeValue > 0
-  ) {
-    return "rsak";
-  }
-
-  if (data.definition.tags.includes("Damage")) {
-    return "msak";
-  }
-
-  if (data.definition.tags.includes("Healing")) {
-    return "heal";
-  }
-
-  if (data.definition.tags.includes("Buff")) {
-    return "util";
-  }
-
-  return "other";
-};
-
-let getAlternativeFormula = (data) => {
-  // this might be specificially for Toll the Dead only, but it's better than nothing
-
-  let description = data.definition.description;
-  let match = description.match(/instead[\w\s]+(\d+d\d+) (\w+) damage/);
-  if (match) {
-    return match[1];
-  } else {
-    return "";
-  }
-};
-
-let getDamage = (data) => {
-  let result = {
-    parts: [],
-    versatile: "",
-  };
-
-  // damage
-  const attacks = data.definition.modifiers.filter((mod) => mod.type === "damage");
-  if (attacks.length !== 0) {
-    const cantripBoost = data.definition.level === 0 && !!data.flags.vtta.dndbeyond.cantripBoost;
-    attacks.forEach((attack) => {
-      let diceString =
-        attack.usePrimaryStat || cantripBoost ? `${attack.die.diceString} + @mod` : attack.die.diceString;
-      result.parts.push([diceString, attack.subType]);
-    });
-
-    // This is probably just for Toll the dead.
-    const alternativeFormula = getAlternativeFormula(data);
-    result.versatile = cantripBoost ? `${alternativeFormula} + @mod` : alternativeFormula;
-    return result;
-  }
-
-  // healing
-  const heals = data.definition.modifiers.filter((mod) => mod.type === "bonus" && mod.subType === "hit-points");
-  if (heals.length !== 0) {
-    heals.forEach((heal) => {
-      let diceString = heal.usePrimaryStat ? `${heal.die.diceString} + @mod` : heal.die.diceString;
-      result.parts.push([diceString, "healing"]);
-    });
-    return result;
-  }
-  return result;
-};
-
-let getSave = (data) => {
-  if (data.definition.requiresSavingThrow && data.definition.saveDcAbilityId) {
-    const saveAbility = DICTIONARY.character.abilities.find((ability) => ability.id === data.definition.saveDcAbilityId)
-      .value;
-    if (data.overrideSaveDc) {
-      return {
-        ability: saveAbility,
-        dc: data.overrideSaveDc,
-        scaling: "flat",
-      };
-    } else {
-      return {
-        ability: saveAbility,
-        dc: null,
-        scaling: "spell",
-      };
-    }
-  } else {
-    return {
-      ability: "",
-      dc: null,
-    };
-  }
-};
-
-/**
- * Get the scaling type for a spell mod
- * This is complex logic and is broken out to help simplify
- * @param {*} name
- * @param {*} mod
- */
-let getScaleType = (name, mod) => {
-  // scaleTypes:
-  // SPELLSCALE - typical spells that scale
-  // SPELLLEVEL - these spells have benefits that come in at particular levels e.g. bestow curse, hex. typically  duration changes
-  // CHARACTERLEVEL - typical cantrip based levelling, some expections (eldritch blast)
-  let scaleType = null;
-  const modScaleType = mod.atHigherLevels.scaleType;
-  const isHigherLevelDefinitions =
-    mod.atHigherLevels.higherLevelDefinitions &&
-    Array.isArray(mod.atHigherLevels.higherLevelDefinitions) &&
-    mod.atHigherLevels.higherLevelDefinitions.length >= 1;
-
-  if (isHigherLevelDefinitions && modScaleType === "spellscale") {
-    const definition = mod.atHigherLevels.higherLevelDefinitions[0];
-    if (definition) {
-      scaleType = modScaleType;
-    } else {
-      utils.log("No definition found for " + name); // eslint-disable-line no-console
-    }
-  } else if (modScaleType === "spellscale") {
-    // lets handle cases where there is a spellscale type but no damage
-    // increase/ higherleveldefinitins e.g. chain lighting
-    // these type of spells typically increase targets so we set the
-    // scaling to null as we don't want to increase damage when upcast.
-    // this also deals with cases like Ice Knife where the upscale damage
-    // is in one of the two mods provided.
-    // we are capturing this else because we don't want to trigger
-    // an update to scaleType or a warning.
-  } else if (modScaleType === "characterlevel") {
-    // lets handle odd cantrips like Eldritch Blast
-    // (in fact this might be the only case)
-    if (mod.atHigherLevels.higherLevelDefinitions.length === 0) {
-      // if this array is empty it does not contain levelling information
-      // the only case found is Eldritch Blast.
-      // this does have some info around multiple beams in
-      // data.atHigherLevels but we ignore this. we will set the scaling
-      // to null as each beam is best modelled by "casting" the cantrip again/
-      // pressing the attack/damage buttons in FVTT
-      scaleType = null;
-    } else {
-      scaleType = modScaleType;
-    }
-  } else if (modScaleType === "spelllevel") {
-    // spells that have particular level associated benefits
-    // these seem to be duration increases or target increases for
-    // the most part we can't handle these in FVTT right now (we could
-    // in theory create a new spell at a higher level).
-    // some duration upcasting (like bestow curse) affects concentration
-    // for now we will do nothing with these spells.
-    // examples include: hex, shadowblade, magic weapon, bestow curse
-    scaleType = modScaleType;
-  } else {
-    utils.log(name + " parse failed: " + JSON.stringify(modScaleType));
-    scaleType = modScaleType; // if this is new/unknow will use default
-  }
-
-  return scaleType;
-};
-
-let getSpellScaling = (data) => {
-  let baseDamage = "";
-  let scaleDamage = "";
-  let scaleType = null; // defaults to null, so will be picked up as a None scaling spell.
-
-  // spell scaling
-  if (data.definition.canCastAtHigherLevel) {
-    // iterate over each spell modifier
-    data.definition.modifiers
-      .filter((mod) => mod.type === "damage" || (mod.type === "bonus" && mod.subType === "hit-points"))
-      .forEach((mod) => {
-        // if the modifier has a die for damage, lets use the string or fixed value
-        // for the base damage
-        if (mod && mod.die) {
-          if (mod.die.diceString !== null) {
-            baseDamage = mod.die.diceString;
-          }
-
-          if (mod.die.fixedValue !== null && baseDamage === "") {
-            baseDamage = mod.die.fixedValue;
-          }
-        }
-
-        // defines some details about higher level casting
-        if (mod.atHigherLevels) {
-          // scaleTypes:
-          // SPELLSCALE - typical spells that scale
-          // SPELLLEVEL - these spells have benefits that come in at particular levels e.g. bestow curse, hex. typically  duration changes
-          // CHARACTERLEVEL - typical cantrip based levelling, some expections (eldritch blast)
-
-          // mod.atHigherLevels.higherLevelDefinitions contains info about the
-          // spells damage die at higher levels, but we can't use this for cantrips as
-          // FVTT use a formula to work out the scaling (ddb has a fixed value structure)
-          const isHigherLevelDefinitions =
-            mod.atHigherLevels.higherLevelDefinitions &&
-            Array.isArray(mod.atHigherLevels.higherLevelDefinitions) &&
-            mod.atHigherLevels.higherLevelDefinitions.length >= 1;
-
-          // lets handle normal spell leveling first
-          if (isHigherLevelDefinitions && mod.atHigherLevels.scaleType === "spellscale") {
-            const definition = mod.atHigherLevels.higherLevelDefinitions[0];
-            if (definition) {
-              const modScaleDamage =
-                definition.dice && definition.dice.diceString // if dice string
-                  ? definition.dice.diceString // use dice string
-                  : definition.dice && definition.dice.fixedValue // else if fixed value
-                  ? definition.dice.fixedValue // use fixed value
-                  : definition.value; // else use value
-
-              // some spells have multiple scaling damage (e.g. Wall of Ice,
-              // Glyph of warding, Acid Arrow, Arcane Hand, Dragon's Breath,
-              // Chromatic Orb, Absorb Elements, Storm Sphere, Spirit Guardians)
-              // it's hard to model most of these in FVTT, and for some it makes
-              // no difference. so...
-              // lets optimistically use the highest
-              // assumptions: these are going to be dice strings, and we don't care
-              // about dice value, just number of dice
-              const diceFormula = /(\d*)d\d*/;
-              const existingMatch = diceFormula.exec(scaleDamage);
-              const modMatch = diceFormula.exec(modScaleDamage);
-
-              if (!existingMatch || modMatch[1] > existingMatch[1]) {
-                scaleDamage = modScaleDamage;
-              }
-            } else {
-              console.warn("No definition found for " + data.definition.name); // eslint-disable-line no-console
-            }
-          }
-
-          scaleType = getScaleType(data.definition.name, mod);
-        }
-      });
-  }
-
-  switch (scaleType) {
-    case "characterlevel":
-      return {
-        mode: "cantrip",
-        formula: "",
-      };
-    case "spellscale":
-      return {
-        mode: "level",
-        formula: scaleDamage,
-      };
-    case "spelllevel":
-    case null:
-      return {
-        mode: "none",
-        formula: "",
-      };
-    default:
-      return {
-        mode: "level",
-        formula: "",
-      };
-  }
-};
-
-// is there a spell casting ability?
-let hasSpellCastingAbility = (spellCastingAbilityId) => {
-  return DICTIONARY.character.abilities.find((ability) => ability.id === spellCastingAbilityId) !== undefined;
-};
-
-// convert spellcasting ability id to string used by vtta
-let convertSpellCastingAbilityId = (spellCastingAbilityId) => {
-  return DICTIONARY.character.abilities.find((ability) => ability.id === spellCastingAbilityId).value;
-};
-
-// search through classinfo and determine spellcasting ability
-let getSpellCastingAbility = (classInfo) => {
-  let spellCastingAbility = undefined;
-  if (hasSpellCastingAbility(classInfo.definition.spellCastingAbilityId)) {
-    spellCastingAbility = convertSpellCastingAbilityId(classInfo.definition.spellCastingAbilityId);
-  } else if (
-    classInfo.subclassDefinition &&
-    hasSpellCastingAbility(classInfo.subclassDefinition.spellCastingAbilityId)
-  ) {
-    // Arcane Trickster has spellcasting ID granted here
-    spellCastingAbility = convertSpellCastingAbilityId(classInfo.subclassDefinition.spellCastingAbilityId);
-  } else {
-    // special cases: No spellcaster, but can cast spells like totem barbarian, default to wis
-    spellCastingAbility = "wis";
-  }
-  return spellCastingAbility;
-};
-
-let getEldritchInvocations = (data) => {
-  let damage = 0;
-  let range = 0;
-
-  const eldritchBlastMods = data.character.modifiers.class.filter(
-    (modifier) => modifier.type === "eldritch-blast" && modifier.isGranted
-  );
-
-  eldritchBlastMods.forEach((mod) => {
-    switch (mod.subType) {
-      case "bonus-damage": {
-        // almost certainly CHA :D
-        const abilityModifier = DICTIONARY.character.abilities.find((ability) => ability.id === mod.statId).value;
-        damage = `@abilities.${abilityModifier}.mod`;
-        break;
-      }
-      case "bonus-range":
-        range = mod.value;
-        break;
-      default:
-        console.warn(`Not yet able to process ${mod.subType}, please raise an issue.`); // eslint-disable-line no-console
-    }
-  });
-
-  return {
-    damage: damage,
-    range: range,
-  };
-};
-
-/**
- * Some spells we need to fix up or massage because they are modified
- * in interesting ways
- * @param {*} ddb
- * @param {*} items
- */
-let fixSpells = (ddb, items) => {
-  items.forEach((spell) => {
-    switch (spell.name) {
-      // Eldritch Blast is a special little kitten and has some fun Eldritch
-      // Invocations which can adjust it.
-      case "Eldritch Blast": {
-        const eldritchBlastMods = getEldritchInvocations(ddb);
-        spell.data.damage.parts[0][0] += " + " + eldritchBlastMods["damage"];
-        spell.data.range.value += eldritchBlastMods["range"];
-        break;
-      }
-      // The target/range input data are incorrect on some AOE spells centreted
-      // on self.
-      // Range is self with an AoE target of 15 ft cube
-      // i.e. affects all creatures within 5 ft of caster
-      case "Thunderclap":
-      case "Word of Radiance":
-        spell.data.range = { value: null, units: "self", long: null };
-        spell.data.target = { value: "15", units: "ft", type: "cube" };
-        break;
-      case "Sleep": {
-        spell.data.damage = { parts: [["5d8", ""]], versatile: "", value: "" };
-        spell.data.scaling = { mode: "level", formula: "2d8" };
-        break;
-      }
-      case "Color Spray": {
-        spell.data.damage = { parts: [["6d10", ""]], versatile: "", value: "" };
-        spell.data.scaling = { mode: "level", formula: "2d10" };
-        break;
-      }
-      // no default
-    }
-  });
-};
-
-let getLookups = (character) => {
-  // racialTraits
-  let lookups = {
-    race: [],
-    feat: [],
-    class: [],
-    classFeature: [],
-    item: [],
-  };
-  character.race.racialTraits.forEach((trait) => {
-    lookups.race.push({
-      id: trait.definition.id,
-      name: trait.definition.name,
-    });
-  });
-
-  character.classes.forEach((playerClass) => {
-    lookups.class.push({
-      id: playerClass.definition.id,
-      name: playerClass.definition.name,
-    });
-
-    if (playerClass.subclassDefinition) {
-      lookups.class.push({
-        id: playerClass.subclassDefinition.id,
-        name: playerClass.subclassDefinition.name,
-      });
-    }
-
-    if (playerClass.classFeatures) {
-      playerClass.classFeatures.forEach((trait) => {
-        lookups.classFeature.push({
-          id: trait.definition.id,
-          name: trait.definition.name,
-          componentId: trait.definition.componentId,
-        });
-      });
-    }
-  });
-
-  character.options.class.forEach((trait) => {
-    lookups.classFeature.push({
-      id: trait.definition.id,
-      name: trait.definition.name,
-      componentId: trait.componentId,
-    });
-  });
-
-  character.feats.forEach((trait) => {
-    lookups.feat.push({
-      id: trait.definition.id,
-      name: trait.definition.name,
-      componentId: trait.componentId,
-    });
-  });
-
-  character.inventory.forEach((trait) => {
-    lookups.item.push({
-      id: trait.definition.id,
-      name: trait.definition.name,
-      limitedUse: trait.limitedUse,
-      equipped: trait.equipped,
-      isAttuned: trait.isAttuned,
-      canAttune: trait.definition.canAttune,
-      canEquip: trait.definition.canEquip,
-    });
-  });
-
-  return lookups;
-};
-
-let parseSpell = (data, character) => {
-  /**
-   * MAIN parseSpell
-   */
-  let spell = {
-    type: "spell",
-    data: JSON.parse(utils.getTemplate("spell")),
-    flags: {
-      vtta: {
-        dndbeyond: data.flags.vtta.dndbeyond,
-      },
-    },
-  };
-
-  // spell name
-  if (data.flags.vtta.dndbeyond.nameOverride !== undefined) {
-    spell.name = data.flags.vtta.dndbeyond.nameOverride;
-  } else {
-    spell.name = data.definition.name;
-  }
-
-  // add tags
-  spell.flags.vtta.dndbeyond.tags = data.definition.tags;
-
-  // spell level
-  spell.data.level = data.definition.level;
-
-  // get the spell school
-  spell.data.school = utils.findInConfig("spellSchools", data.definition.school);
-
-  /**
-   * Gets the necessary spell components VSM + material
-   */
-  spell.data.components = getComponents(data);
-
-  spell.data.materials = getMaterials(data);
-
-  spell.data.preparation = getSpellPreparationMode(data);
-
-  spell.data.description = {
-    value: data.definition.description,
-    chat: data.definition.description,
-    unidentified: data.definition.type,
-  };
-
-  spell.data.source = utils.parseSource(data.definition);
-
-  spell.data.activation = getActivation(data);
-
-  spell.data.duration = getDuration(data);
-
-  spell.data.target = getTarget(data);
-
-  spell.data.range = getRange(data);
-
-  spell.data.actionType = getActionType(data);
-
-  spell.data.damage = getDamage(data);
-
-  spell.data.save = getSave(data);
-
-  spell.data.scaling = getSpellScaling(data);
-
-  spell.data.uses = getUses(data);
-
-  // attach the spell ability id to the spell data so VTT always uses the
-  // correct one, useful if multi-classing and spells have different
-  // casting abilities
-  if (character.data.attributes.spellcasting !== data.flags.vtta.dndbeyond.ability) {
-    spell.data.ability = data.flags.vtta.dndbeyond.ability;
-  }
-
-  // If using better rolls we set alt to be versatile for spells like
-  // Toll The Dead
-  spell.flags.betterRolls5e = {
-    quickVersatile: {
-      altValue: true,
-    },
-  };
-
-  return spell;
-};
-
-export default function parseSpells(ddb, character) {
-  let items = [];
-  const proficiencyModifier = character.data.attributes.prof;
-  const lookups = getLookups(ddb.character);
-
-  // each class has an entry here, each entry has spells
-  // we loop through each class and process
-  ddb.character.classSpells.forEach((playerClass) => {
-    const classInfo = ddb.character.classes.find((cls) => cls.id === playerClass.characterClassId);
-    const spellCastingAbility = getSpellCastingAbility(classInfo);
-    const abilityModifier = utils.calculateModifier(character.data.abilities[spellCastingAbility].value);
-
-    const cantripBoost =
-      ddb.character.modifiers.class.filter(
-        (mod) =>
-          mod.type === "bonus" &&
-          mod.subType === `${classInfo.definition.name.toLowerCase()}-cantrip-damage` &&
-          (mod.restriction === null || mod.restriction === "")
-      ).length > 0;
-
-    // parse spells chosen as spellcasting (playerClass.spells)
-    playerClass.spells.forEach((spell) => {
-      if (!spell.definition) return;
-      // add some data for the parsing of the spells into the data structure
-      spell.flags = {
-        vtta: {
-          dndbeyond: {
-            lookup: "classSpell",
-            class: classInfo.definition.name,
-            level: classInfo.level,
-            spellLevel: spell.definition.level,
-            spellSlots: character.data.spells,
-            ability: spellCastingAbility,
-            mod: abilityModifier,
-            dc: 8 + proficiencyModifier + abilityModifier,
-            cantripBoost: cantripBoost,
-            overrideDC: false,
-          },
-        },
-      };
-
-      // Check for duplicate spells, normally domain ones
-      // We will import spells from a different class that are the same though
-      // as they may come from with different spell casting mods
-      const duplicateSpell = items.findIndex(
-        (existingSpell) =>
-          existingSpell.name === spell.definition.name &&
-          classInfo.definition.name === existingSpell.flags.vtta.dndbeyond.class
-      );
-      if (!items[duplicateSpell]) {
-        items.push(parseSpell(spell, character));
-      } else if (spell.alwaysPrepared) {
-        // if our new spell is always known we overwrite!
-        // it's probably domain
-        items[duplicateSpell] = parseSpell(spell, character);
-      } else {
-        // we'll emit a console message if it doesn't match this case for future debugging
-        console.warn(`Duplicate Spell ${spell.definition.name} detected in class ${classInfo.definition.name}.`); // eslint-disable-line no-console
-      }
-    });
-  });
-
-  // Parse any spells granted by class features, such as Barbarian Totem
-  ddb.character.spells.class.forEach((spell) => {
-    if (!spell.definition) return;
-    // If the spell has an ability attached, use that
-    let spellCastingAbility = undefined;
-    const classInfo = lookups.classFeature.find((cls) => cls.id === spell.componentId);
-    const klass = utils.getClassFromOptionID(ddb, spell.componentId);
-
-    if (hasSpellCastingAbility(spell.spellCastingAbilityId)) {
-      spellCastingAbility = convertSpellCastingAbilityId(spell.spellCastingAbilityId);
-    } else if (klass) {
-      spellCastingAbility = getSpellCastingAbility(klass);
-      // force these spells to always be prepared
-      spell.alwaysPrepared = true;
-    } else {
-      // if there is no ability on spell, we default to wis
-      spellCastingAbility = "wis";
-    }
-
-    const abilityModifier = utils.calculateModifier(character.data.abilities[spellCastingAbility].value);
-
-    // add some data for the parsing of the spells into the data structure
-    spell.flags = {
-      vtta: {
-        dndbeyond: {
-          lookup: "classFeature",
-          lookupName: classInfo.name,
-          lookupId: classInfo.id,
-          level: character.flags.vtta.dndbeyond.totalLevels,
-          ability: spellCastingAbility,
-          mod: abilityModifier,
-          dc: 8 + proficiencyModifier + abilityModifier,
-          overrideDC: false,
-        },
-      },
-    };
-
-    // Check for duplicate spells, normally domain ones
-    // We will import spells from a different class that are the same though
-    // as they may come from with different spell casting mods
-    const duplicateSpell = items.findIndex(
-      (existingSpell) =>
-        existingSpell.name === spell.definition.name &&
-        klass &&
-        klass.definition.name === existingSpell.flags.vtta.dndbeyond.class
-    );
-    if (!items[duplicateSpell]) {
-      items.push(parseSpell(spell, character));
-    } else if (spell.alwaysPrepared) {
-      // if our new spell is always known we overwrite!
-      // it's probably domain
-      items[duplicateSpell] = parseSpell(spell, character);
-    } else {
-      // we'll emit a console message if it doesn't match this case for future debugging
-      utils.log(`Duplicate Spell ${spell.definition.name} detected in class ${classInfo.definition.name}.`);
-    }
-  });
-
-  // Race spells are handled slightly differently
-  ddb.character.spells.race.forEach((spell) => {
-    if (!spell.definition) return;
-    // for race spells the spell spellCastingAbilityId is on the spell
-    // if there is no ability on spell, we default to wis
-    let spellCastingAbility = "wis";
-    if (hasSpellCastingAbility(spell.spellCastingAbilityId)) {
-      spellCastingAbility = convertSpellCastingAbilityId(spell.spellCastingAbilityId);
-    }
-
-    const abilityModifier = utils.calculateModifier(character.data.abilities[spellCastingAbility].value);
-
-    let raceInfo = lookups.race.find((rc) => rc.id === spell.componentId);
-
-    if (!raceInfo) {
-      // for some reason we haven't matched the race option id with the spell
-      // this happens with at least the SCAG optional spells casting half elf
-      raceInfo = {
-        name: "Racial spell",
-        id: spell.componentId,
-      };
-    }
-
-    // add some data for the parsing of the spells into the data structure
-    spell.flags = {
-      vtta: {
-        dndbeyond: {
-          lookup: "race",
-          lookupName: raceInfo.name,
-          lookupId: raceInfo.id,
-          race: ddb.character.race.fullName,
-          level: spell.castAtLevel,
-          ability: spellCastingAbility,
-          mod: abilityModifier,
-          dc: 8 + proficiencyModifier + abilityModifier,
-          overrideDC: false,
-        },
-      },
-    };
-
-    items.push(parseSpell(spell, character));
-  });
-
-  // feat spells are handled slightly differently
-  ddb.character.spells.feat.forEach((spell) => {
-    if (!spell.definition) return;
-    // If the spell has an ability attached, use that
-    // if there is no ability on spell, we default to wis
-    let spellCastingAbility = "wis";
-    if (hasSpellCastingAbility(spell.spellCastingAbilityId)) {
-      spellCastingAbility = convertSpellCastingAbilityId(spell.spellCastingAbilityId);
-    }
-
-    const abilityModifier = utils.calculateModifier(character.data.abilities[spellCastingAbility].value);
-
-    let featInfo = lookups.feat.find((ft) => ft.id === spell.componentId);
-
-    if (!featInfo) {
-      // for some reason we haven't matched the feat option id with the spell
-      // we fiddle the result
-      featInfo = {
-        name: "Feat option spell",
-        id: spell.componentId,
-      };
-    }
-
-    // add some data for the parsing of the spells into the data structure
-    spell.flags = {
-      vtta: {
-        dndbeyond: {
-          lookup: "feat",
-          lookupName: featInfo.name,
-          lookupId: featInfo.id,
-          level: spell.castAtLevel,
-          ability: spellCastingAbility,
-          mod: abilityModifier,
-          dc: 8 + proficiencyModifier + abilityModifier,
-          overrideDC: false,
-        },
-      },
-    };
-
-    items.push(parseSpell(spell, character));
-  });
-
-  if (items) fixSpells(ddb, items);
-
-  return items;
-}
-
-export function parseItemSpells(ddb, character) {
-  let items = [];
-  const proficiencyModifier = character.data.attributes.prof;
-  const lookups = getLookups(ddb.character);
-
-  // feat spells are handled slightly differently
-  ddb.character.spells.item.forEach((spell) => {
-    if (!spell.definition) return;
-    const itemInfo = lookups.item.find((it) => it.id === spell.componentId);
-
-    const active =
-      (!itemInfo.canEquip && !itemInfo.canAttune) || // if item just gives a thing
-      itemInfo.isAttuned || // if it is attuned (assume equipped)
-      (!itemInfo.canAttune && itemInfo.equipped); // can't attune but is equipped
-    // for item spells the spell dc is often on the item spell
-    let spellDC = 8;
-    if (spell.overrideSaveDc) {
-      spellDC = spell.overrideSaveDc;
-    } else if (spell.spellCastingAbilityId) {
-      // If the spell has an ability attached, use that
-      // if there is no ability on spell, we default to wis
-      let spellCastingAbility = "wis";
-      if (hasSpellCastingAbility(spell.spellCastingAbilityId)) {
-        spellCastingAbility = convertSpellCastingAbilityId(spell.spellCastingAbilityId);
-      }
-
-      const abilityModifier = utils.calculateModifier(character.data.abilities[spellCastingAbility].value);
-      spellDC = 8 + proficiencyModifier + abilityModifier;
-    } else {
-      spellDC = null;
-    }
-
-    // add some data for the parsing of the spells into the data structure
-    spell.flags = {
-      vtta: {
-        dndbeyond: {
-          lookup: "item",
-          lookupName: itemInfo.name,
-          lookupId: itemInfo.id,
-          level: spell.castAtLevel,
-          dc: spellDC,
-          limitedUse: itemInfo.limitedUse,
-          nameOverride: `${spell.definition.name} (${itemInfo.name})`,
-          overrideDC: !!spell.overrideSaveDc,
-          spellLimitedUse: spell.limitedUse,
-          castAtLevel: spell.castAtLevel,
-          active: active,
-        },
-      },
-    };
-
-    items.push(parseSpell(spell, character));
-  });
-
-  if (items) fixSpells(ddb, items);
-
-  return items;
-}
